Fixed error in safe_unique_values returning none if data was not yet loaded, it should return []
