+++ callbacks/callbacks.py
@@ -37,9 +37,10 @@ def register_callbacks(app, data_manager: DataManager):
         Output('project-selector', 'options')],
         [Input('refresh-data', 'n_clicks')],
         [State('project-filter', 'options'),
-        State('employee-filter', 'options')]
+        State('employee-filter', 'options'),
+        State('project-selector', 'options')]
     )
-    def refresh_dashboard_data(n_clicks, current_project_options, current_employee_options):
+    def refresh_dashboard_data(n_clicks, current_portfolio_options, current_employee_options, current_project_options):
         logger.info(f"refresh_dashboard_data called. n_clicks: {n_clicks}")
         ctx = dash.callback_context
         if not ctx.triggered:
@@ -58,13 +59,14 @@ def register_callbacks(app, data_manager: DataManager):
             df_projects = data_manager.df_portfolio
             df_employees = data_manager.df_employees
 
-            project_options = [{'label': i, 'value': i} for i in df_projects['name'].unique() if pd.notna(i)]
+            portfolio_options = [{'label': i, 'value': i} for i in df_projects['name'].unique() if pd.notna(i)]
             employee_options = [{'label': i, 'value': i} for i in df_employees['name'].unique() if pd.notna(i)]
+            project_options = portfolio_options # same as portfolio but only one can be chosen
 
-            return last_update, project_options, employee_options, project_options
+            return last_update, portfolio_options, employee_options, project_options
         else:
             logger.warning("Data is empty")
-            return "Failed to update data", current_project_options, current_employee_options, current_project_options
+            return "Failed to update data", current_portfolio_options, current_employee_options, current_project_options
 
     @app.callback(
         Output('project-filter', 'disabled'),
@@ -91,36 +93,36 @@ def register_callbacks(app, data_manager: DataManager):
                 date_column = date_columns[0]
             else:
                 return go.Figure()  # Return empty figure if no suitable date column found
-        
+
         filtered_sales = data_manager.df_sales[
             (data_manager.df_sales[date_column] >= start_date) &
             (data_manager.df_sales[date_column] <= end_date)
         ]
-        
+
         filtered_tasks = data_manager.df_tasks[
             (data_manager.df_tasks['create_date'] >= start_date) &
             (data_manager.df_tasks['create_date'] <= end_date)
         ]
-        
+
         if task_filter:
             keywords = [keyword.strip().lower() for keyword in task_filter.split(',')]
             filtered_tasks = filtered_tasks[filtered_tasks['name'].str.lower().str.contains('|'.join(keywords))]
-        
+
         if filtered_sales.empty and filtered_tasks.empty:
             return go.Figure()
-        
+
         daily_sales = filtered_sales.groupby(date_column)['amount_total'].sum().reset_index()
         daily_tasks = filtered_tasks.groupby('create_date').size().reset_index(name='task_count')
-        
+
         fig = go.Figure()
         fig.add_trace(go.Scatter(x=daily_sales[date_column], y=daily_sales['amount_total'], name='Sales', mode='lines'))
         fig.add_trace(go.Scatter(x=daily_tasks['create_date'], y=daily_tasks['task_count'], name='Tasks', mode='lines', yaxis='y2'))
-        
+
         fig.update_layout(
             title='Sales and Tasks Over Time',
             xaxis_title='Date',
             yaxis_title='Sales Amount',
             yaxis2=dict(title='Number of Tasks', overlaying='y', side='right')
         )
-        
+
         return fig
