+++ cfg/version.txt
@@ -1,1 +1,1 @@
-24.08.06.11.37
+24.08.07.12.52
